fixed bug on validations

+++ app/scripts/controllers/login.controller.js
@@ -20,27 +20,38 @@ app.controller('LoginCtrl',['$scope',
   };
 
   $scope.login = function(params) {
-    dataFactory.getConfirm().then(function(response) {
-      if(response.data.confirmed || !!$scope.emailConfirmedByUserOnce) {
-        $scope.emailConfirmedByUserOnce = true;
-        authFactory.login(params).success(function(response) {
-          $window.sessionStorage.setItem('OnMyPlate.user', response.token);
-          // Sets the headers for the request, and token for the authorization
-          $http.defaults.headers.common['Authorization'] = 'Token token=' + $window.sessionStorage.getItem('OnMyPlate.user');
-          $location.path('/');
-        }).error(function(response) {
-          $scope.params = {};
-          $scope.isLoginSuccessful = false;
-          $('#login-error').slideDown(200);
-          $('#login-error').delay(3000).slideUp(200);
+    dataFactory.fetchUsers().then(function(response) {
+      var doesUserExist = response.data.users.filter(function(user) {return user.email === params.email})[0];
+      if(!!doesUserExist) {
+        dataFactory.getConfirm().then(function(response) {
+          if(response.data.confirmed || !!$scope.emailConfirmedByUserOnce) {
+            $scope.emailConfirmedByUserOnce = true;
+            authFactory.login(params).success(function(response) {
+              $window.sessionStorage.setItem('OnMyPlate.user', response.token);
+              // Sets the headers for the request, and token for the authorization
+              $http.defaults.headers.common['Authorization'] = 'Token token=' + $window.sessionStorage.getItem('OnMyPlate.user');
+              $location.path('/');
+            }).error(function(response) {
+              $scope.params = {};
+              $scope.isLoginSuccessful = false;
+              $('#login-error').slideDown(200);
+              $('#login-error').delay(3000).slideUp(200);
+            });
+          } else {
+            $location.path('/login');
+            $scope.isConfirmed = false;
+            $('#confirmation-error').slideDown(200);
+            $('#confirmation-error').delay(3000).slideUp(200);
+          }
         });
       } else {
-        $location.path('/login');
-        $scope.isConfirmed = false;
-        $('#confirmation-error').slideDown(200);
-        $('#confirmation-error').delay(3000).slideUp(200);
+        $scope.params = {};
+        $scope.isLoginSuccessful = false;
+        $('#login-error').slideDown(200);
+        $('#login-error').delay(3000).slideUp(200);
       }
     });
+    
   }
 
 }]);
